+++ src/main/java/com/bawnorton/trulyrandom/extend/BrewingStandBlockEntityExtender.java
@@ -1,0 +1,9 @@
+package com.bawnorton.trulyrandom.extend;
+
+import net.minecraft.world.item.crafting.RecipeHolder;
+
+import java.util.Map;
+
+public interface BrewingStandBlockEntityExtender {
+    Map<Integer, RecipeHolder<?>> trulyrandom$RECIPES();
+}