+++ src/main/java/com/bawnorton/trulyrandom/mixin/recipe/BrewingRecipeMixin.java
@@ -1,39 +1,27 @@
+//? if >=26.3 {
 package com.bawnorton.trulyrandom.mixin.recipe;
 
 import com.bawnorton.trulyrandom.extend.ResultClearer;
 import com.llamalad7.mixinextras.injector.ModifyReturnValue;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.crafting.*;
+import net.minecraft.world.item.crafting.BrewingRecipe;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 
-@Mixin({
-        DyeRecipe.class,
-        BannerDuplicateRecipe.class,
-        BookCloningRecipe.class,
-        DecoratedPotRecipe.class,
-        FireworkRocketRecipe.class,
-        FireworkStarFadeRecipe.class,
-        FireworkStarRecipe.class,
-        MapExtendingRecipe.class,
-        RepairItemRecipe.class,
-        ShieldDecorationRecipe.class,
-        ImbueRecipe.class,
-        SmithingTrimRecipe.class,
-        TransmuteRecipe.class
-})
-abstract class SpecialRecipeMixin implements ResultClearer {
+@Mixin(BrewingRecipe.class)
+abstract class BrewingRecipeMixin implements ResultClearer {
     @Unique
     private ItemStack result = ItemStack.EMPTY;
 
     @ModifyReturnValue(
-            method = "assemble(Lnet/minecraft/world/item/crafting/RecipeInput;)Lnet/minecraft/world/item/ItemStack;",
+            method = "assemble(Lnet/minecraft/world/item/crafting/BrewingInput;)Lnet/minecraft/world/item/ItemStack;",
             at = @At("RETURN")
     )
     private ItemStack useRandomResult(ItemStack result) {
         if (this.result.isEmpty()) return result;
-        return this.result;
+
+        return this.result.copy();
     }
 
     @Override
@@ -46,3 +34,4 @@ abstract class SpecialRecipeMixin implements ResultClearer {
         return result;
     }
 }
+//?}